updated / renamed some paths
